# apps/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Portfolio tracking API",
        "title": "My Portfolio API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/holdings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holdings"
                ],
                "summary": "List holdings with pagination",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default 20, max 100)",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/PaginatedHoldings"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holdings"
                ],
                "summary": "Create a new holding",
                "parameters": [
                    {
                        "description": "Holding payload",
                        "name": "holding",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateHoldingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Holding"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/holdings/{id}": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "holdings"
                ],
                "summary": "Update an existing holding",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Holding ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Holding payload",
                        "name": "holding",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateHoldingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Holding"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "CreateHoldingRequest": {
            "type": "object",
            "required": [
                "asset_type",
                "description"
            ],
            "properties": {
                "asset_type": {
                    "type": "string"
                },
                "average_cost_basis": {
                    "type": "number"
                },
                "cost_basis_total": {
                    "type": "number"
                },
                "current_value": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "dividend_income": {
                    "type": "number"
                },
                "issuer_jurisdiction": {
                    "type": "string"
                },
                "last_price": {
                    "type": "number"
                },
                "purchase_quantity": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "tax_class_override": {
                    "description": "TaxClassOverride and IssuerJurisdiction describe the asset's tax\nattributes; nil means \"derive from the asset type\". They replace the old\nstate_tax_exempt flag - what a holding pays and who issued it are facts\nabout the asset, while whether that's exempt is a jurisdiction's rule.",
                    "type": "string"
                }
            }
        },
...
        "Holding": {
            "type": "object",
            "properties": {
                "asset_type": {
                    "type": "string"
                },
                "average_cost_basis": {
                    "type": "number"
                },
                "cost_basis_total": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "current_value": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "dividend_income": {
                    "type": "number"
                },
                "gain_realized_amount": {
                    "type": "number"
                },
                "gain_realized_percent": {
                    "type": "number"
                },
                "gain_unrealized_amount": {
                    "type": "number"
                },
                "gain_unrealized_percent": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "issuer_jurisdiction": {
                    "description": "IssuerJurisdiction is the government that issued this asset's debt, for\nthe bond classes. It's what the tax rules compare against a jurisdiction\nto decide questions like \"did my own state issue this?\" - nil means the\nasset-type default (see defaultIssuerJurisdiction).",
                    "type": "string"
                },
                "last_price": {
                    "type": "number"
                },
                "purchase_quantity": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "tax_class": {
                    "description": "TaxClass is the resolved class (override, else the asset-type default),\ncomputed for responses via AfterFind and never persisted.",
                    "type": "string"
                },
                "tax_class_override": {
                    "description": "TaxClassOverride overrides the tax class implied by the asset type. Nil\nmeans \"derive from the asset type\" (see ResolveTaxClass); a non-nil value\nis an explicit user choice - e.g. a treasury-only ETF like TLT classed as\ngovernment_bond even though ETFs are equity by default.",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "PaginatedHoldings": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Holding"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "UpdateHoldingRequest": {
            "type": "object",
            "properties": {
                "asset_type": {
                    "type": "string"
                },
                "average_cost_basis": {
                    "type": "number"
                },
                "cost_basis_total": {
                    "type": "number"
                },
                "current_value": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "dividend_income": {
                    "type": "number"
                },
                "issuer_jurisdiction": {
                    "type": "string"
                },
                "last_price": {
                    "type": "number"
                },
                "purchase_quantity": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "tax_class_override": {
                    "type": "string"
                }
            }
        },
...
    }
}